Adding unit tests to a new file
